Add coalesced tests
